Refactoring GenericResource -> Resource
